backend/requirements.txt
anyio==4.13.0
bcrypt==5.0.0
google-auth==2.40.0
certifi==2026.2.25
click==8.3.2
fastapi==0.136.0
h11==0.16.0
httpx==0.28.1
idna==3.11
pydantic==2.13.1
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
python-multipart==0.0.26
pywebpush==2.0.0
SQLAlchemy==2.0.49
starlette==1.0.0
urllib3==2.6.3
uvicorn[standard]==0.44.0
